BLTI-259 BLTI-254
127985 BLTI-259
127990, 127998 BLTI-254



git-svn-id: https://source.sakaiproject.org/svn/basiclti/branches/basiclti-2.1.x@128964 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ basiclti-impl/src/java/org/sakaiproject/lti/impl/DBLTIService.java
@@ -134,9 +134,9 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 	 * 
 	 * @see org.sakaiproject.lti.api.LTIService#getToolDao(java.lang.Long, java.lang.String, boolean)
 	 */
-	protected Map<String, Object> getToolDao(Long key, String siteId, boolean isAdminRole, boolean isMaintainRole) 
+	protected Map<String, Object> getToolDao(Long key, String siteId, boolean isAdminRole) 
 	{
-		return getThingDao("lti_tools", LTIService.TOOL_MODEL, key, siteId, isAdminRole, isMaintainRole);
+		return getThingDao("lti_tools", LTIService.TOOL_MODEL, key, siteId, isAdminRole);
 	}
 
 	/**
@@ -168,8 +168,8 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 	 *      int, int, java.lang.String, boolean)
 	 */
 	public List<Map<String, Object>> getToolsDao(String search, String order, int first,
-			int last, String siteId, boolean isAdminRole, boolean isMaintainRole) {
-		return getThingsDao("lti_tools", LTIService.TOOL_MODEL, search, order, first, last, siteId, isAdminRole, isMaintainRole);
+			int last, String siteId, boolean isAdminRole) {
+		return getThingsDao("lti_tools", LTIService.TOOL_MODEL, search, order, first, last, siteId, isAdminRole);
 	}
 
 	/**
@@ -178,6 +178,14 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 	protected Object insertContentDao(Properties newProps, String siteId, 
 		boolean isAdminRole, boolean isMaintainRole) 
 	{
+		if ( newProps == null ) {
+			throw new IllegalArgumentException(
+					"newProps must be non-null");
+		}
+		if (siteId == null && !isAdminRole ) {
+			throw new IllegalArgumentException("siteId must be non-null for non-admins");
+		}
+
 		String toolId = newProps.getProperty(LTIService.LTI_TOOL_ID);
 		if (toolId == null)
 			return rb.getString("error.missing.toolid");
@@ -190,7 +198,7 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 
 		// Load the tool we are aiming for Using DAO
 		Map<String, Object> tool = null;
-		tool = getToolDao(toolKey, siteId, isAdminRole, isMaintainRole);
+		tool = getToolDao(toolKey, siteId, isAdminRole);
 			
 		if ( tool == null ) {
 			return rb.getString("error.invalid.toolid");
@@ -204,7 +212,7 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 			}
 		}
 
-		String[] contentModel = getContentModelDao(tool, siteId, isAdminRole, isMaintainRole);
+		String[] contentModel = getContentModelDao(tool, isAdminRole);
 		String[] columns = foorm.getFields(contentModel);
 		
 		// Since page title and title are both required and dynamically hideable, 
@@ -244,8 +252,8 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 	 * 
 	 * @see org.sakaiproject.lti.api.LTIService#getContentDao(java.lang.Long, java.lang.String, boolean)
 	 */
-	public Map<String, Object> getContentDao(Long key, String siteId, boolean isAdminRole, boolean isMaintainRole) {
-		Map<String, Object> retval = getThingDao("lti_content", LTIService.CONTENT_MODEL, key, siteId, isAdminRole, isMaintainRole);
+	public Map<String, Object> getContentDao(Long key, String siteId, boolean isAdminRole) {
+		Map<String, Object> retval = getThingDao("lti_content", LTIService.CONTENT_MODEL, key, siteId, isAdminRole);
 		if (retval == null) return retval;
 		retval.put("launch_url", getContentLaunch(retval));
 		return retval;
@@ -269,9 +277,19 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 	 * @see org.sakaiproject.lti.impl.BaseLTIService#updateContentDao(java.lang.Long, 
 	 *      java.lang.Object, java.lang.String, boolean)
 	 */
-	public Object updateContentDao(Long key, Object newProps, String siteId, boolean isAdminRole, boolean isMaintainRole) {
+	public Object updateContentDao(Long key, Object newProps, String siteId, 
+		boolean isAdminRole, boolean isMaintainRole) 
+	{
+		if ( key == null || newProps == null ) {
+			throw new IllegalArgumentException(
+					"both key and newProps must be non-null");
+		}
+		if (siteId == null && !isAdminRole ) {
+			throw new IllegalArgumentException("siteId must be non-null for non-admins");
+		}
+
 		// Load the content item
-		Map<String,Object> content = getContentDao(key, siteId, isAdminRole, isMaintainRole);
+		Map<String,Object> content = getContentDao(key, siteId, isAdminRole);
 		if (  content == null ) {
 			return rb.getString("error.content.not.found");
 		}
@@ -291,7 +309,7 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 		if ( newToolKey == null || newToolKey < 0 ) newToolKey = oldToolKey;
 
 		// Load the tool we are aiming for
-		Map<String, Object> tool = getToolDao(newToolKey, siteId, isAdminRole, isMaintainRole);
+		Map<String, Object> tool = getToolDao(newToolKey, siteId, isAdminRole);
 		if ( tool == null ) {
 			return rb.getString("error.invalid.toolid");
 		}
@@ -306,7 +324,7 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 			}
 		}
 
-		String[] contentModel = getContentModelDao(tool, siteId, isAdminRole, isMaintainRole);
+		String[] contentModel = getContentModelDao(tool, isAdminRole);
 		if (contentModel == null)
 			return rb.getString("error.invalid.toolid");
 
@@ -322,9 +340,9 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 	 *      java.lang.String, int, int)
 	 */
 	public List<Map<String, Object>> getContentsDao(String search, String order, int first,
-			int last, String siteId, boolean isAdminRole, boolean isMaintainRole) {
+			int last, String siteId, boolean isAdminRole) {
 		List<Map<String, Object>> contents = getThingsDao("lti_content",
-				LTIService.CONTENT_MODEL, search, order, first, last, siteId, isAdminRole, isMaintainRole);
+				LTIService.CONTENT_MODEL, search, order, first, last, siteId, isAdminRole);
 		for (Map<String, Object> content : contents) {
 			content.put("launch_url", getContentLaunch(content));
 		}
@@ -338,9 +356,12 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 	public Object insertThingDao(String table, String[] formModel, String[] fullModel,
 			Properties newProps, String siteId, boolean isAdminRole, boolean isMaintainRole) {
 
-		if (table == null || formModel == null || newProps == null || siteId == null) {
+		if (table == null || formModel == null || newProps == null ) {
 			throw new IllegalArgumentException(
-					"siteId, table, model, and newProps must all be non-null");
+					"table, model, and newProps must all be non-null");
+		}
+		if (siteId == null && !isAdminRole ) {
+			throw new IllegalArgumentException("siteId must be non-null for non-admins");
 		}
 
 		if (!isMaintainRole) return null;
@@ -360,10 +381,10 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 			columns = foorm.getFields(fullModel);
 			theKey = foorm.formSqlKey(fullModel);
 		}
-		if ((Arrays.asList(columns).indexOf(LTIService.LTI_SITE_ID) >= 0)) {
-			if (!isAdminRole && newMapping.get(LTIService.LTI_SITE_ID) == null) {
-				newMapping.put(LTIService.LTI_SITE_ID, siteId);
-			}
+
+		// Only admins can insert things into sites other than the current site
+		if (!isAdminRole && (Arrays.asList(columns).indexOf(LTIService.LTI_SITE_ID) >= 0)) {
+			newMapping.put(LTIService.LTI_SITE_ID, siteId);
 		}
 		String seqName = foorm.getSqlSequence(table, theKey, m_sql.getVendor());
 
@@ -433,17 +454,21 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 	 * @return
 	 */
 	private Map<String, Object> getThingDao(String table, String[] model, Long key,
-			String siteId, boolean isAdminRole, boolean isMaintainRole)
+			String siteId, boolean isAdminRole)
 	{
 		if (table == null || model == null || key == null) {
 			throw new IllegalArgumentException("table, model, and key must all be non-null");
+		}
+		if (siteId == null && !isAdminRole ) {
+			throw new IllegalArgumentException("siteId must be non-null for non-admins");
 		}
 		String statement = "SELECT " + foorm.formSelect(model) + " from " + table
 			+ " WHERE id = ?";
 		Object fields[] = null;
 		String[] columns = foorm.getFields(model);
 
-		if (siteId != null && Arrays.asList(columns).indexOf(LTIService.LTI_SITE_ID) >= 0 && !isAdminRole ) {
+		// Non-admins only see global (SITE_ID IS NULL) or in their site
+		if (!isAdminRole && Arrays.asList(columns).indexOf(LTIService.LTI_SITE_ID) >= 0 ) {
 			statement += " AND (SITE_ID = ? OR SITE_ID IS NULL)";
 			fields = new Object[2];
 			fields[0] = key;
@@ -474,14 +499,20 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 	 * @return
 	 */
 	public List<Map<String, Object>> getThingsDao(String table, String[] model, String search,
-			String order, int first, int last, String siteId, boolean isAdminRole, boolean isMaintainRole) {
+		String order, int first, int last, String siteId, boolean isAdminRole) 
+	{
 		if (table == null || model == null ) {
 			throw new IllegalArgumentException("table and model must be non-null");
 		}
+		if (siteId == null && !isAdminRole ) {
+			throw new IllegalArgumentException("siteId must be non-null for non-admins");
+		}
+
 		String statement = "SELECT " + foorm.formSelect(model) + " FROM " + table;
 		String[] columns = foorm.getFields(model);
 		String whereClause = "";
 
+		// Only admins can see invisible items and items from any site
 		Object fields[] = null;
 		if ( ! isAdminRole ) {
 			if (Arrays.asList(columns).indexOf(LTIService.LTI_VISIBLE) >= 0 && 
@@ -523,14 +554,20 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 		if (table == null || model == null || key == null) {
 			throw new IllegalArgumentException("table, model, and key must all be non-null");
 		}
+		if (siteId == null && !isAdminRole ) {
+			throw new IllegalArgumentException("siteId must be non-null for non-admins");
+		}
+
+		if (!isMaintainRole) return false;
+
 		String statement = "DELETE FROM " + table + " WHERE id = ?";
 		Object fields[] = null;
 		String[] columns = foorm.getFields(model);
 
-		// Hack to insure that We *Can* delete this since SqlService cannot tell us if updates
-		// work
+		// TODO: Check if this is needed on all three databases now that KNL-775 is in place
+		// Hack to insure that we *can* delete this since SqlService cannot tell us if deletes work
 		if (!isAdminRole) {
-			Object thing = getThingDao(table, model, key, siteId, isAdminRole, isMaintainRole);
+			Object thing = getThingDao(table, model, key, siteId, isAdminRole);
 			if (thing == null || !(thing instanceof Map)) {
 				return false;
 			}
@@ -542,11 +579,8 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 			}
 		}
 
-		if (Arrays.asList(columns).indexOf(LTIService.LTI_SITE_ID) >= 0 && !isAdminRole ) {
-			if (!isMaintainRole) {
-				M_log.info("Non-maintain attemped delete on " + table);
-				return false;
-			}
+		// Only admins can delete by id irrespective of the current site
+		if (!isAdminRole && Arrays.asList(columns).indexOf(LTIService.LTI_SITE_ID) >= 0 ) {
 			statement += " AND SITE_ID = ?";
 			fields = new Object[2];
 			fields[0] = key;
@@ -594,8 +628,11 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 			throw new IllegalArgumentException(
 					"table, model, key, and newProps must all be non-null");
 		}
+		if (siteId == null && !isAdminRole ) {
+			throw new IllegalArgumentException("siteId must be non-null for non-admins");
+		}
 
-		if (!isMaintainRole) return null;
+		if (!isMaintainRole) return false;
 
 		HashMap<String, Object> newMapping = new HashMap<String, Object>();
 
@@ -610,10 +647,9 @@ public class DBLTIService extends BaseLTIService implements LTIService {
 			columns = foorm.getFields(fullModel);
 		}
 
-		if ( (Arrays.asList(columns).indexOf(LTIService.LTI_SITE_ID) >= 0)) {
-			if ( !isAdminRole && newMapping.get(LTIService.LTI_SITE_ID) == null) {
-				newMapping.put(LTIService.LTI_SITE_ID, siteId);
-			}
+		// Only admins can update *into* a site
+		if ( !isAdminRole && (Arrays.asList(columns).indexOf(LTIService.LTI_SITE_ID) >= 0)) {
+			newMapping.put(LTIService.LTI_SITE_ID, siteId);
 		}
 
 		String sql = "UPDATE " + table + " SET " + foorm.updateForm(newMapping)
